+++ spring-xsuaa/src/test/java/com/sap/cloud/security/xsuaa/XsuaaServiceConfigurationCustomTest.java
@@ -23,7 +23,7 @@ class XsuaaServiceConfigurationCustomTest {
 		credentials.setXsAppName("xsappname");
 		credentials.setCertificate("-----BEGIN CERTIFICATE-----");
 		credentials.setCertUrl("cert-url");
-		credentials.setPrivateKey("-----BEGIN RSA PRIVATE KEY-----");
+		credentials.setKey("-----BEGIN RSA PRIVATE KEY-----");
 		credentials.setCredentialType(CredentialType.X509);
 
 		cut = new XsuaaServiceConfigurationCustom(credentials);
@@ -39,7 +39,7 @@ class XsuaaServiceConfigurationCustomTest {
 		assertEquals(credentials.getCredentialType(), cut.getCredentialType());
 		assertEquals(credentials.getCertUrl(), cut.getCertUrl().toString());
 		assertEquals(credentials.getCertificate(), cut.getClientIdentity().getCertificate());
-		assertEquals(credentials.getPrivateKey(), cut.getClientIdentity().getKey());
+		assertEquals(credentials.getKey(), cut.getClientIdentity().getKey());
 	}
 
 	@Test